Full project commit :star:
The COVID-19 pandemic caused many “office workers” to move their activity to their homes. Now when most severe cases are rare, home-working remains very popular. Companies and institutions are not fully prepared to handle this challenging circumstance. One of the problems that arise is the monitoring of workers’ activity. The goal of the proposed application is to supply functionality that would allow reporting workers’ activity without additional effort. The system should be as simple as possible.